remove dialog tree
